+++ frontend/drivers/platform_ps3.c
@@ -632,5 +632,7 @@ frontend_ctx_driver_t frontend_ctx_ps3 = {
    NULL,                         /* destroy_sighandler_state */
    NULL,                         /* attach_console */
    NULL,                         /* detach_console */
+   NULL,                         /* watch_path_for_changes */
+   NULL,                         /* check_for_path_changes */
    "ps3",
 };