+++ menu/menu_displaylist.c
@@ -1367,15 +1367,18 @@ static int menu_displaylist_parse_shader_options(menu_displaylist_info_t *info)
    unsigned pass_count = menu_shader_manager_get_amount_passes();
 
    menu_entries_append_enum(info->list,
-         msg_hash_to_str(MENU_ENUM_LABEL_VALUE_SHADER_PLACEHOLDER),
-         msg_hash_to_str(MENU_ENUM_LABEL_SHADER_PLACEHOLDER),
-         MENU_ENUM_LABEL_SHADER_PLACEHOLDER,
-         MENU_SETTING_ACTION, 0, 0);
-   menu_entries_append_enum(info->list,
          msg_hash_to_str(MENU_ENUM_LABEL_VALUE_SHADER_APPLY_CHANGES),
          msg_hash_to_str(MENU_ENUM_LABEL_SHADER_APPLY_CHANGES),
          MENU_ENUM_LABEL_SHADER_APPLY_CHANGES,
          MENU_SETTING_ACTION, 0, 0);
+   if (frontend_driver_can_watch_for_changes())
+   {
+      menu_entries_append_enum(info->list,
+            msg_hash_to_str(MENU_ENUM_LABEL_VALUE_SHADER_WATCH_FOR_CHANGES),
+            msg_hash_to_str(MENU_ENUM_LABEL_SHADER_WATCH_FOR_CHANGES),
+            MENU_ENUM_LABEL_SHADER_WATCH_FOR_CHANGES,
+            0, 0, 0);
+   }
    menu_entries_append_enum(info->list,
          msg_hash_to_str(MENU_ENUM_LABEL_VALUE_VIDEO_SHADER_PRESET),
          msg_hash_to_str(MENU_ENUM_LABEL_VIDEO_SHADER_PRESET),